Performance : Removing `HibernateTemplate` usage to control overhead (#1059)
Due to the change, `EventHandler` (and possibly other classes) can
be remarkably simplified through using the session directly. However,
it's not clear if all the functionality (flush, etc.) from the
template is covered. Needs more testing.


git-svn-id: file:///home/svn/omero/trunk@3747 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/server/src/ome/services/sharing/FreezeShareStore.java
@@ -66,22 +66,22 @@ public class FreezeShareStore extends ShareStore {
     // =========================================================================
 
     @Override
-    public int totalShares() {
+    public Long totalShares() {
         ShareMap map = new ShareMap(conn, envMap, false);
         try {
             int mapsize = map.size();
-            return mapsize;
+            return Long.valueOf(mapsize);
         } finally {
             map.close();
         }
     }
 
     @Override
-    public int totalSharedItems() {
+    public Long totalSharedItems() {
         ShareItems items = new ShareItems(conn, envItems, false);
         try {
             int itemssize = items.size();
-            return itemssize;
+            return Long.valueOf(itemssize);
         } finally {
             items.close();
         }
